本地储存实现+异常处理完善

+++ common/src/main/java/com/LAB/study/dto/PictureDTO.java
@@ -5,10 +5,10 @@ import java.time.LocalDateTime;
 
 @Data
 public class PictureDTO {
-    private Integer pictureid;
+    private Long pictureid;
     private String fileName;
     private LocalDateTime shotTime;
-    private String fileUrl; // 原始链接或基础链接
-    private String previewUrl; // 动态生成的七牛云预览/缩略图链接
+    private String fileUrl;
+    private String previewUrl;
     private String fileExif;
 }
